[pred-memopt] Change AvailableValue to be its own struct instead of a typealias to std::pair.
For pred-memopt to work with ownership, we need to insert instructions in two
different places:

1. When loading not-available values, we must insert the load at the site of the
load we are trying to promote.
2. When propagating an available value, we must destructure and or copy before
each one of the stores that provide our value.

By changing AvailableValue to be a struct, I am providing a cleaner API, but
more importantly the ability to start tracking that information.

rdar://31521023