SILGen: Modernize transparent_attribute test
When the test was written we had a 'transparent' attribute on the apply SIL
instruction. Now that this is gone, most of the test was not testing anything
useful. Remove most of the test, get it to actually test that transparent
function's default argument generators are transparent, and move the test
for @_versioned to the versioned_attribute.swift test.