[pred-memopt] Extract out dataflow computation into its own helper struct.
This simplifies the code and enables me to extract out the code for deleting
dead allocations into a separate struct from AllocOptimize. Otherwise, I would
have to pass in a reference to AllocOptimize creating back references, a
complexity that isn't necessary and will force me to break encapsulation
properties.

Just chopping this off of a larger commit.

rdar://31521023