Standardize indentation of Convention code.
In some places we were using:

namespace {
  class Foo {
  };
}

and in other places we were doing:

namespace {
class Foo {
};
}

This commit standardizes on the later since it is "correct" LLVM style. I am
fixing this in preparation for my work in this area to add support for always
using +0 when a convention is implicit.

rdar://34222540