[test] Add some tests for non-deserializable generic requirements
The existing code already handles this, so it's just a matter of
testing what happens for these declarations that can be generic:
functions, initializers, and subscripts. (Enums are coming in the next
commit.) Note that we currently don't try to recover at all for a
struct or class whose generic requirements aren't valid anymore.