[silgen] Add a new convention for allocating inits.
In an allocating init, we forward our inputs to the underlying initializing
init. Since initializing inits take that value at +1, the allocating init should
as well to avoid unnecessary copies. Previously, this was accomplished by having
the allocating init just use the default argument convention (i.e. +1). But in a
world where "normal" arguments are passed at +0, that doesn't work anymore. So
in this commit, I add a new convention just for allocating inits that takes
"normal" arguments at +0 and asserts if asked for self.

rdar://34222540