[silgen] Teach emitApplyOfLibraryIntrinsic how to convert arguments to the appropriate +0/+1 convention.
Sometimes we need to copy/borrow now.

*NOTE* right now it is basically impossible to test this using a fake stdlib
since I would need to implement Array or String in my local stdlib since this
comes up via ArrayToPointer/friends. If/when we flip the switch though, this is
guaranteed to be tested by stdlib dependent tests that we already have in tree.

rdar://34222540