Turn off ownership checking for this test.
When this branch was cut, SIL ownership generation and checking were
still under development, so it's not a given that a particular pattern
of code will satisfy it.  It makes more sense to just turn off ownership
checking in this test on the branch than to fix all the problems that
would allow it to stay on.