[sil] Catch if we are creating ABI incompatible convert_function instructions when constructing the inst.
I ran into this earlier today. It is a pretty annoying error to track down in
SILGen since the verifier doesn't run until the full function is emitted. With
this change, we get a check immediately when the value is constructed by
SILBuilder making it a very easy problem to track down.