build-script-impl: Partial implementation for a "quick" xctoolchain
Building an xctoolchain for use in Xcode is hard.
utils/build-toolchain makes it easy, but it's not so great for
actively working on swiftc. Add support to build-script-impl
(hopefully to be exposed in a pretty way elsewhere) for building a
fake xctoolchain out of symlinks.

Only swiftc/SourceKit and LLDB are supported right now; someone else
can do the work for playgrounds.

To play with this, try these commands with build-script:
  --darwin-toolchain-build-symlink
  --install-destdir=$PWD
  --toolchain-prefix=test.xctoolchain
  --darwin_toolchain_bundle_identifier=org.swift.MASTER
  --darwin-toolchain-display-name "Swift master"