[silgen] Split out long logic in emitEpilogBB into a helper function so we can use early returns to simplify the logic.
In a subsequent commit, I am going to be changing emitEpilogBB to use more
modern SILGen to ensure that the return value from the epilogBB has a proper
cleanup. I was able to create a case where that happened.

Slicing this off of a larger commit to ease review.

rdar://34222540