IRGen: Redo fill ops
Instead of recording the current offset in the template, just
collect the fill ops into a vector and emit them at the
generic argument offset stored in the MetadataLayout.

For now this is NFC, but this allows the template to omit
entries for generic arguments entirely once some more
machinery is in place.