Set SIL stage to lowered at the *start* of address lowering
SIL is not in a consistent state during the lowering pass. We just make sure it is correct at the end of the pass. This broke PR https://github.com/apple/swift/pull/13283 (verifying convert_function instructions in the builder)

To workaround this issue we will be running the builder’s verifier when the SIL stage != SILStage::Lowered

The SIL stage is set to lowered at the end of the module pass. This means that if this is the first lowering pass said workaround would not work.

This commit sets the SIL stage to lowered at the beginning of the module pass instead